Consolidate validated security and robustness fixes (#115)
* Validate FastSort input bounds

* Fix deque index wraparound overflow

* Harden Solution2 stdin handling

* Avoid recursion in BST deletion

* Handle malformed calculator expressions

* Handle empty input in Distances

* Harden AverageElement input processing

* Handle blank input in EncryptThis

* Prevent RecursionMax stack exhaustion

* Fail smoke verification on any invocation error

* Prevent overflow in sliding average updates

* Fix sequence counting across separated zeros

* Handle literal characters in CamelCase conversion

* Handle zero in expanded form

* fix: bound tribonacci sequence allocation

* Handle null input in Maskify

* Validate Rail Fence cipher rail counts

* Handle empty Pig Latin tokens safely

* Harden Distances input and output handling

* Harden rail fence cipher decoding

+++ src/main/java/algorithms/sprint0/SlidingAverage.java
@@ -28,7 +28,7 @@ public class SlidingAverage {
         }
         result.add(sum / (double) windowSize);
         for (int i = windowSize; i < minSize; i++) {
-            sum += arr.get(i) - arr.get(i - windowSize);
+            sum += (long) arr.get(i) - arr.get(i - windowSize);
             result.add(sum / (double) windowSize);
         }
         return result;
